api: round out dev helpers for template emails
