+++ .github/workflows/ci.yml
@@ -82,8 +82,8 @@ jobs:
       - uses: actions/cache@v4
         with:
           path: api/.build
-          key: api-${{ hashFiles('api/Package.resolved') }}
-          restore-keys: api-
+          key: api-build-${{ hashFiles('api/Package.resolved') }}
+          restore-keys: api-build-
       - name: meta
         id: meta
         uses: friends-library/dev/actions/ts-pack/actions/meta@master
@@ -152,8 +152,8 @@ jobs:
       - uses: actions/cache/restore@v4
         with:
           path: api/.build
-          key: api-${{ hashFiles('api/Package.resolved') }}
-          restore-keys: api-
+          key: api-test-${{ hashFiles('api/Package.resolved') }}
+          restore-keys: api-test-
       - name: test-api
         run: cd api && swift test
 
@@ -211,7 +211,7 @@ env:
   DASHBOARD_URL: /dashboard
   ANALYTICS_SITE_URL: /analytics
   POSTMARK_API_KEY: not-real
-  POSTMARK_SERVER_ID: not-real
+  POSTMARK_SERVER_ID: 12345
   PRIMARY_SUPPORT_EMAIL: not-real
   SUPER_ADMIN_EMAIL: not-real
   STRIPE_SUBSCRIPTION_PRICE_ID: not-real
